refactor: rules module
