+++ package.json
@@ -98,6 +98,7 @@
   },
   "files": [
     "lib/index.js",
-    "lib/index.d.ts"
+    "lib/index.d.ts",
+    "lib/rules.js"
   ]
 }
